move tables to tables.js
